refactor(infrastructure): MarkdownTaskAdapterをMarkdownTaskClientに名称変更
- adapters/からclients/へ配置変更
- クラス名をMarkdownTaskClient に変更（Clientとして明確化）
- 関連ドキュメント（MVP_PLAN.md, AGENTS.md）を更新
- adapters/index.tsを空のモジュールに変更（将来のAdapter用）
