+++ src/infrastructure/adapters/index.ts
@@ -1,11 +1,3 @@
 // Infrastructure Adapters
-export {
-	type CreateTaskInfo,
-	type FrontmatterConfig,
-	MarkdownParseError,
-	MarkdownTaskAdapter,
-	type ParsedTask,
-	type ParseResult,
-	SerializerError,
-	type TaskEdit,
-} from './markdownTaskAdapter';
+// 現在はアダプターはありません。将来、MarkdownTaskRepositoryなどが追加される予定です。
+export {};
