+++ src/infrastructure/clients/index.ts
@@ -1,2 +1,13 @@
 // Infrastructure Clients
+
+export {
+	type CreateTaskInfo,
+	type FrontmatterConfig,
+	MarkdownParseError,
+	MarkdownTaskClient,
+	type ParsedTask,
+	type ParseResult,
+	SerializerError,
+	type TaskEdit,
+} from './markdownTaskClient';
 export { type FrontmatterResult, RemarkClient } from './remarkClient';
